# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "Your API Description",
        "version": "1.0.0"
    },
    "paths": {
        "/customers/{id}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get a customer by ID",
                "operationId": "2e4adaa335a3b883549d8abe8c2f694c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Customer ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Customer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Customers"
                ],
                "summary": "Update a customer",
                "operationId": "846f265d18cd9adfc7381bbb1ffe7d2d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Customer ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Customers"
                ],
                "summary": "Delete a customer",
                "operationId": "0a43a81f6fbfd4a16f41ccddd5a78314",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Customer ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/customers": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get all customers",
                "operationId": "00a03daeab0d5be3e9471b68b0f96817",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Customer"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Create a new customer",
                "operationId": "77311275857c095c590f704140af9718",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CustomerStoreRequest": {
                "title": "Customer Store Request",
                "description": "Customer store request body",
                "required": [
                    "first_name",
                    "date_of_birth",
                    "last_name",
                    "email",
                    "phone_number",
                    "bank_account_number"
                ],
                "properties": {
                    "first_name": {
                        "type": "string",
                        "maxLength": 40,
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "maxLength": 40,
                        "example": "Doe"
                    },
                    "date_of_birth": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-01-01"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 150,
                        "example": "johndoe@example.com"
                    },
                    "phone_number": {
                        "type": "string",
                        "maxLength": 25,
                        "example": "+1234567890"
                    },
                    "bank_account_number": {
                        "type": "string",
                        "maxLength": 36,
                        "example": "1234567890"
                    }
                },
                "type": "object"
            },
            "CustomerUpdateRequest": {
                "title": "Customer Update Request",
                "description": "Customer update request body",
                "required": [
                    "first_name",
                    "date_of_birth",
                    "last_name",
                    "email",
                    "phone_number",
                    "bank_account_number"
                ],
                "properties": {
                    "first_name": {
                        "type": "string",
                        "maxLength": 40,
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "maxLength": 40,
                        "example": "Doe"
                    },
                    "date_of_birth": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-01-01"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 150,
                        "example": "johndoe@example.com"
                    },
                    "phone_number": {
                        "type": "string",
                        "maxLength": 25,
                        "example": "+1234567890"
                    },
                    "bank_account_number": {
                        "type": "string",
                        "maxLength": 36,
                        "example": "1234567890"
                    }
                },
                "type": "object"
            },
            "Customer": {
                "title": "Customer",
                "description": "Customer model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "first_name": {
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "date_of_birth": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-01-01"
                    },
                    "email": {
                        "type": "string",
                        "example": "johndoe@example.com"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "bank_account_number": {
                        "type": "string",
                        "example": "1234567890"
                    }
                },
                "type": "object"
            }
        }
    }
}